lambda-functions/new_project/requirements.txt
boto3==1.35.71
python-gitlab==5.1.0
requests==2.32.3
